Multiple RNN Cells and Layers (#358)
* Added multiple RNN cell/layer functionality

* Finished multiple rnn functionality

* Modified rnn example and finished rnn testing

* init

* examples cuda fixes

* skip large conversion

* change log

* pycodestyle
