+++ CHANGELOG.md
@@ -19,11 +19,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 * Set weights can be used to re-apply the weight scaling omega. (\#360)
 * Out scaling factors can be learnt even if weight scaling omega was set to 0. (\#360)
 * Reverse up / down option for ``LinearStepDevice``. (\#361)
+* Generic Analog RNN classes (LSTM, RNN, GRU) uni or bidirectional. (\#358)
 
 ### Fixed
 
 * Legacy checkpoint load with alpha scaling. (\#360)
-* Re-application of weight scaling omega when loading checkpoints. (\#360) 
+* Re-application of weight scaling omega when loading checkpoints. (\#360)
 
 ### Changed
 
@@ -82,7 +83,7 @@ The format is based on [Keep a Changelog], and this project adheres to
   types if set larger than zero. (\#318)
 * The use of generators for analog tiles of an ``AnalogModuleBase``. (\#320)
 * Digital bias is now accessable through ``MappingParameter``. (\#331)
-* The aihwkit documentation. New content around analog ai concepts, training presets, analog ai 
+* The aihwkit documentation. New content around analog ai concepts, training presets, analog ai
   optimizers, new references, and examples. (\#348)
 * The `weight_scaling_omega` can now be defined in the `rpu_config.mapping`. (\#353)
