+++ src/aihwkit/nn/__init__.py
@@ -17,7 +17,8 @@
 from aihwkit.nn.modules.container import AnalogSequential
 from aihwkit.nn.modules.conv import AnalogConv1d, AnalogConv2d, AnalogConv3d
 from aihwkit.nn.modules.linear import AnalogLinear
-from aihwkit.nn.modules.lstm import AnalogLSTM
+from aihwkit.nn.modules.rnn.rnn import AnalogRNN
+from aihwkit.nn.modules.rnn.cells import AnalogGRUCell, AnalogLSTMCell, AnalogVanillaRNNCell
 from aihwkit.nn.modules.linear_mapped import AnalogLinearMapped
 from aihwkit.nn.modules.conv_mapped import (
     AnalogConv1dMapped, AnalogConv2dMapped, AnalogConv3dMapped
